Refactor calcResponse and prepare for deployment
Added if statement in app.js so morgan is only used
in a development environment.

Added the cors npm package to prepare for a front-end
that is hosted on another server.

Refactored queryStringToDate so only valid dates will
be accepted. Added relevant tests as well.

Created a production folder that will store javascript
that has already been transpiled for use, so babel-node
will not have to be used at run time.
Also added relevant npm scripts for building and serving
the transpiled files.